requirements.txt
numpy
flax
optax
jaxopt
einops
datasets
seaborn
matplotlib
tqdm
scikit-learn
